requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
python-dotenv
httpx
tenacity
